Comment out updating of email address from DfE SignIn
Can't support this till the FK on email as been removed

System.InvalidOperationException   HResult=0x80131509   Message=The property 'Email' on entity type 'McUser' is part of a key and so cannot be modified or marked as modified. To change the principal of an existing entity with an identifying foreign key first delete the dependent and invoke 'SaveChanges' then associate the dependent with the new principal.