Populate email in identity from mc_user instead of sign-in claims
This makes our system's knowledge of the email internally consistent.

This fixes an invalid operation exception caused by the terms filter looking for a user by email address when their sign-in email address didn't match the one in mc_users. (We could still find the user because they had previously signed in with their old address and we had their subjectId logged).