Refactor: Extract mapping logic from DataService and Enrichment service
This splits out the methods LoadCourse and LoadCourses (and dependency methods) into a new CourseLoader class which is used by DataService.

Similarly, translating Enrichment objects to and from JSON is extracted from EnrichmentService and encapsulated in a new EnrichmentConverter class.

To keep the refactor simple the new classes are instantiated in situ.