Export all published courses as a JSON
This is a temporary command line utility to export all ManageCourses courses in a format that can be posted to SearchAndCompare.

It contains an amount of duplication that needs to be removed once the CourseMapper refactor has completed (see the "Duplicated" folder and namespace)

Further, the code touches up the Fees, Salary and ProviderLocation objects that are not currently handled by CourseMapper. This logic needs to be moved into CourseMapper once the refactor is finished.