[1033] Skip OptedIn providers when importing from UCAS
We are doing a staged transition from UCAS being the source of truth to
DfE find being the source of truth for providers & courses.
This patch means that when a provider is opted in to the transition
their data will cease to be pulled in from the UCAS data feed. Without
this any data entered in manage courses would be overwritten on the next
import.