[1033] Cope with opted in accrediting providers
Because the opted-in provider is not upserted, the dictionary used to
hook the course up to its accredited provider did not contain the
matching entry, so an exception was thrown which is logged and
swallowed.

This patch includes combines the upsert cache with the cache of values
from the database so opted in providers will be available.